+++ mago3d-admin/src/main/java/com/gaia3d/controller/PolicyController.java
@@ -54,7 +54,7 @@ public class PolicyController {
 		String defaultProjects = policy.getGeo_data_default_projects();
 		if(defaultProjects != null && !"".equals(defaultProjects)) {
 			String[] projectIds = defaultProjects.split(",");
-			Map<Long, Project> projectMap = CacheManager.getProjectMap();
+			Map<Integer, Project> projectMap = CacheManager.getProjectMap();
 			String defaultProjectsView = "";
 			for(String projectId : projectIds) {
 				Project project = projectMap.get(Long.valueOf(projectId));
