+++ mago3d-admin/src/main/java/com/gaia3d/service/FileService.java
@@ -44,7 +44,7 @@ public interface FileService {
 	 * @param fileInfo
 	 * @return
 	 */
-	public FileInfo insertDataFile(Long projectId, FileInfo fileInfo);
+	public FileInfo insertDataFile(Integer projectId, FileInfo fileInfo);
 	
 	/**
 	 * Data Attribute 등록
