+++ mago3d-core/src/main/java/com/gaia3d/domain/CacheParams.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class CacheParams {
 
 	// project 고유 식별자
-	private Long project_id;
+	private Integer project_id;
 	// 캐시명
 	private CacheName cacheName;
 	// 캐시 유형
